refactor(painter-dom): move fragment context type
